feat: article thumbnail styles
